chore: move screen-shooter from core dir up to root lib dir
