+++ lib/browser/commands/assert-view/index.js
@@ -3,7 +3,7 @@
 const fs = require('fs-extra');
 const _ = require('lodash');
 const Image = require('../../../image');
-const ScreenShooter = require('../../../core/screen-shooter');
+const ScreenShooter = require('../../screen-shooter');
 const temp = require('../../../temp');
 const {getCaptureProcessors} = require('./capture-processors');
 const {getTestContext} = require('../../../utils/mocha');
