+++ lib/browser/screen-shooter/index.js
