+++ lib/browser/screen-shooter/viewport/coord-validator/errors/height-viewport-error.js
