+++ lib/browser/screen-shooter/viewport/coord-validator/errors/offset-viewport-error.js
