+++ lib/browser/screen-shooter/viewport/coord-validator/index.js
