+++ lib/browser/screen-shooter/viewport/index.js
