+++ lib/worker/runner/test-runner/one-time-screenshooter.js
@@ -3,7 +3,7 @@
 const Promise = require('bluebird');
 
 const Image = require('../../../image');
-const ScreenShooter = require('../../../core/screen-shooter');
+const ScreenShooter = require('../../../browser/screen-shooter');
 const logger = require('../../../utils/logger');
 
 module.exports = class OneTimeScreenshooter {
