+++ test/lib/browser/commands/assert-view/index.js
@@ -6,7 +6,7 @@ const fs = require('fs-extra');
 const webdriverio = require('webdriverio');
 const clientBridge = require('lib/browser/client-bridge');
 const Image = require('lib/image');
-const ScreenShooter = require('lib/core/screen-shooter');
+const ScreenShooter = require('lib/browser/screen-shooter');
 const temp = require('lib/temp');
 const AssertViewError = require('lib/browser/commands/assert-view/errors/assert-view-error');
 const ImageDiffError = require('lib/browser/commands/assert-view/errors/image-diff-error');
