+++ test/lib/browser/screen-shooter/index.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const Image = require('lib/image');
-const ScreenShooter = require('lib/core/screen-shooter');
-const Viewport = require('lib/core/screen-shooter/viewport');
+const ScreenShooter = require('lib/browser/screen-shooter');
+const Viewport = require('lib/browser/screen-shooter/viewport');
 
 describe('screen-shooter', () => {
     const sandbox = sinon.sandbox.create();
