+++ test/lib/worker/runner/test-runner/one-time-screenshooter.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 const Image = require('lib/image');
-const ScreenShooter = require('lib/core/screen-shooter');
+const ScreenShooter = require('lib/browser/screen-shooter');
 const OneTimeScreenshooter = require('lib/worker/runner/test-runner/one-time-screenshooter');
 const logger = require('lib/utils/logger');
 const {mkSessionStub_} = require('../../../browser/utils');
